First pass at spliting the breakpoint wirting into its own class
