+++ src/managers/BreakpointManager.ts
@@ -4,11 +4,16 @@ import { orderBy } from 'natural-orderby';
 import { SourceNode } from 'source-map';
 import { DebugProtocol } from 'vscode-debugprotocol';
 import { fileUtils } from '../FileUtils';
-import { Project } from './ProjectManager';
+import { Project, ComponentLibraryProject } from './ProjectManager';
 import { standardizePath as s } from 'roku-deploy';
 import { SourceMapManager } from './SourceMapManager';
 import { LocationManager } from './LocationManager';
 import { util } from '../util';
+import { TelnetAdapter } from '../adapters/TelnetAdapter';
+import { DebugProtocolAdapter } from '../adapters/DebugProtocolAdapter';
+import { Breakpoint } from 'vscode-debugadapter';
+import { EventEmitter } from 'events';
+import { AddBreakpointRequestObject } from '../debugProtocol/Debugger';
 
 export class BreakpointManager {
 
@@ -19,12 +24,40 @@ export class BreakpointManager {
 
     }
 
+    public on(eventName: 'changed', handler: (breakpoints: Breakpoint[]) => void);
+    public on(eventName: string, handler: (payload: any) => void) {
+        this.emitter.on(eventName, handler);
+        return () => {
+            this.emitter?.removeListener(eventName, handler);
+        };
+    }
+
+    private emitter = new EventEmitter();
+
+    private emit(
+        eventName:
+            'changed',
+        data?
+    ) {
+        //emit these events on next tick, otherwise they will be processed immediately which could cause issues
+        setTimeout(() => {
+            //in rare cases, this event is fired after the debugger has closed, so make sure the event emitter still exists
+            if (this.emitter) {
+                this.emitter.emit(eventName, data);
+            }
+        }, 0);
+    }
+
     public launchConfiguration: {
         sourceDirs: string[],
         rootDir: string;
         enableSourceMaps?: boolean
     };
 
+    public supportConditionalBreakpoints: boolean = true;
+    public supportHitConditionalBreakpoints: boolean = true;
+    public supportLogPoints: boolean = true;
+    public supportNormalBreakpoints: boolean = true;
     /**
      * Tell the breakpoint manager that no new breakpoints can be verified
      * (most likely due to the app being launched and roku not supporting dynamic breakpoints)
@@ -39,6 +72,10 @@ export class BreakpointManager {
      */
     private areBreakpointsLocked = false;
 
+    private breakpointsQueue: {
+        [refId: string]: DebugProtocol.SetBreakpointsArguments
+    } = {};
+
     /**
      * A map of breakpoints by what file they were set in.
      * This does not handle any source-to-dest mapping...these breakpoints are stored in the file they were set in.
@@ -47,30 +84,50 @@ export class BreakpointManager {
      */
     private breakpointsByFilePath = {} as { [sourceFilePath: string]: AugmentedSourceBreakpoint[] };
 
-    public static breakpointIdSequence = 1;
+    private breakpointRefIds: { [refId: string]: number } = {};
+
+    public queueBreakpointsForFile(args: DebugProtocol.SetBreakpointsArguments) {
+        let sanitizedPath = this.sanitizeSourceFilePath(args.source.path);
+        this.breakpointsQueue[sanitizedPath] = args;
+        let breakpoints = [];
+        for (let breakpoint of args.breakpoints) {
+            breakpoints.push({
+                /** If true breakpoint could be set (but not necessarily at the desired location). */
+                verified: false,
+                id: this.getBreakpointRefId(this.sanitizeSourceFilePath(sanitizedPath), breakpoint.line),
+                /** The source where the breakpoint is located. */
+                source: args.source,
+                // message: 'device is not in a stopped state',
+                /** The start line of the actual range covered by the breakpoint. */
+                line: breakpoint.line
+            });
+        }
+
+        process.nextTick(() => {
+            this.emit('changed', breakpoints);
+        });
+
+        return breakpoints;
+    }
+
+    public getBreakpointQueue(): { [refId: string]: DebugProtocol.SetBreakpointsArguments } {
+        return this.breakpointsQueue;
+    }
+
+    private getBreakpointRefId(sourcePath: string, line: number) {
+        let key = `${sourcePath}:${line}`;
+        if (!this.breakpointRefIds[key]) {
+            let id = Object.keys(this.breakpointRefIds).length + 1;
+            this.breakpointRefIds[key] = id;
+        }
+        return this.breakpointRefIds[key];
+    }
 
     /**
      * breakpoint lines are 1-based, and columns are zero-based
      */
     public registerBreakpoint(sourceFilePath: string, breakpoint: DebugProtocol.SourceBreakpoint | AugmentedSourceBreakpoint) {
-        sourceFilePath = this.sanitizeSourceFilePath(sourceFilePath);
-        //get the breakpoints array (and optionally initialize it if not set)
-        let breakpointsArray = this.breakpointsByFilePath[sourceFilePath] = this.breakpointsByFilePath[sourceFilePath] ?? [];
-
-        let existingBreakpoint = breakpointsArray.find(x => x.line === breakpoint.line);
-
-        let bp = <AugmentedSourceBreakpoint>Object.assign(existingBreakpoint || {}, breakpoint);
-
-        //set column=0 if the breakpoint is missing that field
-        bp.column = bp.column ?? 0;
-
-        bp.wasAddedBeforeLaunch = bp.wasAddedBeforeLaunch ?? this.areBreakpointsLocked === false;
-
-        //set an id if one does not already exist (used for pushing breakpoints to the client)
-        bp.id = bp.id ?? BreakpointManager.breakpointIdSequence++;
-
-        //any breakpoint set in this function is not hidden
-        bp.isHidden = false;
+        let bp = this.updateBreakpoint(sourceFilePath, breakpoint);
 
         //mark non-supported breakpoint as NOT verified, since we don't support debugging non-brightscript files
         if (!fileUtils.hasAnyExtension(sourceFilePath, ['.brs', '.bs', '.xml'])) {
@@ -98,23 +155,88 @@ export class BreakpointManager {
                 sourceFilePath = lastWorkingPath;
 
             }
-            //new breakpoints will be verified=false, but breakpoints that were removed and then added again should be verified=true
-            if (breakpointsArray.find(x => x.wasAddedBeforeLaunch && x.line === bp.line)) {
-                bp.verified = true;
-                bp.wasAddedBeforeLaunch = true;
-            } else {
-                bp.verified = false;
-                bp.wasAddedBeforeLaunch = false;
-            }
+
+            bp.verified = bp.wasAddedBeforeLaunch;
+        }
+    }
+
+    public updateBreakpoint(sourceFilePath: string, breakpoint: DebugProtocol.SourceBreakpoint | AugmentedSourceBreakpoint): AugmentedSourceBreakpoint {
+        sourceFilePath = this.sanitizeSourceFilePath(sourceFilePath);
+        //get the breakpoints array (and optionally initialize it if not set)
+        let breakpointsArray = this.breakpointsByFilePath[sourceFilePath] = this.breakpointsByFilePath[sourceFilePath] ?? [];
+
+        let existingBreakpoint = breakpointsArray.find(x => x.line === breakpoint.line);
+
+        let bp = <AugmentedSourceBreakpoint>Object.assign(existingBreakpoint || {}, breakpoint);
+
+        //set an id if one does not already exist (used for pushing breakpoints to the client)
+        bp.id = bp.id ?? this.getBreakpointRefId(sourceFilePath, breakpoint.line);
+
+        //set column=0 if the breakpoint is missing that field
+        bp.column = bp.column ?? 0;
+
+        bp.wasAddedBeforeLaunch = bp.wasAddedBeforeLaunch ?? this.areBreakpointsLocked === false;
+
+        bp.protocolBreakpointIds = bp.protocolBreakpointIds ?? [];
+
+        //any breakpoint set in this function is not hidden
+        bp.isHidden = false;
+
+        //new breakpoints will be verified=false, but breakpoints that were removed and then added again should be verified=true
+        if (breakpointsArray.find(x => x.wasAddedBeforeLaunch && x.line === bp.line)) {
+            bp.wasAddedBeforeLaunch = true;
+        } else {
+            bp.wasAddedBeforeLaunch = false;
         }
 
-        //if we already have a breakpoint for this exact line, don't add another one
+         //if we already have a breakpoint for this exact line, don't add another one
         if (breakpointsArray.find(x => x.line === breakpoint.line)) {
             return;
         } else {
             //add the breakpoint to the list
             breakpointsArray.push(bp);
         }
+
+        return bp;
+    }
+
+    public async getBreakpointRequests(path: string, breakpoint: DebugProtocol.SourceBreakpoint,  project: ComponentLibraryProject | Project, fileProtocol: string) {
+        let breakpoints = [];
+        console.log(project);
+        let stagingLocationsResult = await this.locationManager.getStagingLocations(
+            path,
+            breakpoint.line,
+            breakpoint.column,
+            [
+                ...project.sourceDirs,
+                project.rootDir
+            ],
+            project.stagingFolderPath,
+            project.fileMappings
+        );
+
+        for (let stagingLocation of stagingLocationsResult.locations) {
+            let relativeStagingPath = fileUtils.replaceCaseInsensitive(
+                stagingLocation.filePath,
+                fileUtils.standardizePath(
+                    fileUtils.removeTrailingSlash(project.stagingFolderPath) + '/'
+                ),
+                ''
+            );
+
+            if (project instanceof ComponentLibraryProject) {
+                // If this is a Component Library project me need tp make sure we add the post fix to the path
+                relativeStagingPath = project.addFileNamePostfix(relativeStagingPath);
+            }
+
+            let breakpointRequest: AddBreakpointRequestObject = {
+                filePath: fileProtocol + '/' + relativeStagingPath,
+                lineNumber: stagingLocation.lineNumber,
+                hitCount: breakpoint.hitCondition ? parseInt(breakpoint.hitCondition) : 0
+            };
+            breakpoints.push(breakpointRequest);
+        }
+        return breakpoints;
     }
 
     /**
@@ -130,7 +252,7 @@ export class BreakpointManager {
             this.breakpointsByFilePath[sourceFilePath] = this.getBreakpointsForFile(sourceFilePath)
                 .filter(x => x.verified);
 
-            //hide all of the breakpoints (the active ones will be reenabled later in this method)
+            //hide all of the breakpoints (the active ones will be re-enabled later in this method)
             for (let bp of this.breakpointsByFilePath[sourceFilePath]) {
                 bp.isHidden = true;
             }
@@ -145,68 +267,6 @@ export class BreakpointManager {
 
         //get the final list of breakpoints
         return this.getBreakpointsForFile(sourceFilePath);
-    }
-
-    /**
-     * Get a list of all breakpoint tasks that should be performed.
-     * This will also exclude files with breakpoints that are not in scope.
-     */
-    private async getBreakpointWork(project: Project) {
-        let result = {} as {
-            [stagingFilePath: string]: Array<BreakpointWorkItem>
-        };
-
-        //iterate over every file that contains breakpoints
-        for (let sourceFilePath in this.breakpointsByFilePath) {
-            let breakpoints = this.breakpointsByFilePath[sourceFilePath];
-
-            for (let breakpoint of breakpoints) {
-                //get the list of locations in staging that this breakpoint should be written to.
-                //if none are found, then this breakpoint is ignored
-                let stagingLocationsResult = await this.locationManager.getStagingLocations(
-                    sourceFilePath,
-                    breakpoint.line,
-                    breakpoint.column,
-                    [
-                        ...project.sourceDirs,
-                        project.rootDir
-                    ],
-                    project.stagingFolderPath,
-                    project.fileMappings
-                );
-
-                for (let stagingLocation of stagingLocationsResult.locations) {
-                    let relativeStagingPath = fileUtils.replaceCaseInsensitive(
-                        stagingLocation.filePath,
-                        fileUtils.standardizePath(
-                            fileUtils.removeTrailingSlash(project.stagingFolderPath) + '/'
-                        ),
-                        ''
-                    );
-                    let obj: BreakpointWorkItem = {
-                        //add the breakpoint info
-                        ...breakpoint,
-                        //add additional info
-                        sourceFilePath: sourceFilePath,
-                        rootDirFilePath: s`${project.rootDir}/${relativeStagingPath}`,
-                        line: stagingLocation.lineNumber,
-                        column: stagingLocation.columnIndex,
-                        stagingFilePath: stagingLocation.filePath,
-                        type: stagingLocationsResult.type
-                    };
-                    if (!result[stagingLocation.filePath]) {
-                        result[stagingLocation.filePath] = [];
-                    }
-                    result[stagingLocation.filePath].push(obj);
-                }
-            }
-        }
-        //sort every breakpoint by location
-        for (let stagingFilePath in result) {
-            result[stagingFilePath] = this.sortAndRemoveDuplicateBreakpoints(result[stagingFilePath]);
-        }
-
-        return result;
     }
 
     public sortAndRemoveDuplicateBreakpoints<T extends { line: number; column?: number; }>(
@@ -227,162 +287,19 @@ export class BreakpointManager {
     }
 
     /**
-     * Write "stop" lines into source code for each breakpoint of each file in the given project
+     * Get the list of breakpoints for the specified file path, or an empty array
      */
-    public async writeBreakpointsForProject(project: Project) {
-        var breakpointsByStagingFilePath = await this.getBreakpointWork(project);
-
-        let promises = [] as Promise<any>[];
-        for (let stagingFilePath in breakpointsByStagingFilePath) {
-            promises.push(this.writeBreakpointsToFile(stagingFilePath, breakpointsByStagingFilePath[stagingFilePath]));
-        }
-
-        await Promise.all(promises);
-    }
-
-    /**
-     * Write breakpoints to the specified file, and update the sourcemaps to match
-     */
-    private async writeBreakpointsToFile(stagingFilePath: string, breakpoints: BreakpointWorkItem[]) {
-
-        //do not crash if the file doesn't exist
-        if (!await fsExtra.pathExists(stagingFilePath)) {
-            util.log(`Path not found ${stagingFilePath}`);
-            return;
-        }
-
-        //load the file as a string
-        let fileContents = (await fsExtra.readFile(stagingFilePath)).toString();
-
-        let originalFilePath = breakpoints[0].type === 'sourceMap' ?
-            //the calling function will merge this sourcemap into the other existing sourcemap, so just use the same name because it doesn't matter
-            breakpoints[0].rootDirFilePath :
-            //the calling function doesn't have a sourcemap for this file, so we need to point it to the sourceDirs found location (probably rootDir...)
-            breakpoints[0].sourceFilePath;
-
-        let sourceAndMap = this.getSourceAndMapWithBreakpoints(fileContents, originalFilePath, breakpoints);
-
-        //if we got a map file back, write it to the filesystem
-        if (sourceAndMap.map) {
-            let sourceMap = JSON.stringify(sourceAndMap.map);
-            //It's ok to overwrite the file in staging because if the original code provided a source map,
-            //then our LocationManager class will walk the sourcemap chain from staging, to rootDir, and then
-            //on to the original location
-            await fsExtra.writeFile(`${stagingFilePath}.map`, sourceMap);
-            //update the in-memory version of this source map
-            this.sourceMapManager.set(`${stagingFilePath}.map`, sourceMap);
-        }
-
-        //overwrite the file that now has breakpoints injected
-        await fsExtra.writeFile(stagingFilePath, sourceAndMap.code);
-    }
-
-    private bpIndex = 1;
-    public getSourceAndMapWithBreakpoints(fileContents: string, originalFilePath: string, breakpoints: BreakpointWorkItem[]) {
-        let chunks = [] as Array<SourceNode | string>;
-
-        //split the file by newline
-        let lines = eol.split(fileContents);
-        let newline = '\n';
-        for (let originalLineIndex = 0; originalLineIndex < lines.length; originalLineIndex++) {
-            let line = lines[originalLineIndex];
-            //if is final line
-            if (originalLineIndex === lines.length - 1) {
-                newline = '';
-            }
-            //find breakpoints for this line (breakpoint lines are 1-indexed, but our lineIndex is 0-based)
-            let lineBreakpoints = breakpoints.filter(bp => bp.line - 1 === originalLineIndex);
-            //if we have a breakpoint, insert that before the line
-            for (let bp of lineBreakpoints) {
-                let linesForBreakpoint = this.getBreakpointLines(bp, originalFilePath);
-
-                //separate each line for this breakpoint with a newline
-                for (let bpLine of linesForBreakpoint) {
-                    chunks.push(bpLine);
-                    chunks.push('\n');
-                }
-            }
-
-            //add the original code now
-            chunks.push(
-                //sourceNode expects 1-based row indexes
-                new SourceNode(originalLineIndex + 1, 0, originalFilePath, `${line}${newline}`)
-            );
-        }
-
-        let node = new SourceNode(null, null, originalFilePath, chunks);
-
-        //if sourcemaps are disabled, skip sourcemap generation and only generate the code
-        if (this.launchConfiguration?.enableSourceMaps === false) {
-            return {
-                code: node.toString(),
-                map: undefined
-            };
-        } else {
-            return node.toStringWithSourceMap();
-        }
-    }
-
-    private getBreakpointLines(breakpoint: BreakpointWorkItem, originalFilePath: string) {
-        let lines = [];
-        if (breakpoint.logMessage) {
-            let logMessage = breakpoint.logMessage;
-            //wrap the log message in quotes
-            logMessage = `"${logMessage}"`;
-            let expressionsCheck = /\{(.*?)\}/g;
-            let match;
-
-            // Get all the value to evaluate as expressions
-            while (match = expressionsCheck.exec(logMessage)) {
-                logMessage = logMessage.replace(match[0], `"; ${match[1]};"`);
-            }
-
-            // add a PRINT statement right before this line with the formated log message
-            lines.push(new SourceNode(breakpoint.line, 0, originalFilePath, `PRINT ${logMessage}`));
-        } else if (breakpoint.condition) {
-            // add a conditional STOP statement
-            lines.push(new SourceNode(breakpoint.line, 0, originalFilePath, `if ${breakpoint.condition} then : STOP : end if`));
-        } else if (breakpoint.hitCondition) {
-            let hitCondition = parseInt(breakpoint.hitCondition);
-
-            if (isNaN(hitCondition) || hitCondition === 0) {
-                // add a STOP statement right before this line
-                lines.push(`STOP`);
-            } else {
-                let prefix = `m.vscode_bp`;
-                let bpName = `bp${this.bpIndex++}`;
-                let checkHits = `if ${prefix}.${bpName} >= ${hitCondition} then STOP`;
-                let increment = `${prefix}.${bpName} ++`;
-
-                // Create the BrightScript code required to track the number of executions
-                let trackingExpression = `
-                    if Invalid = ${prefix} OR Invalid = ${prefix}.${bpName} then
-                        if Invalid = ${prefix} then
-                            ${prefix} = {${bpName}: 0}
-                        else
-                            ${prefix}.${bpName} = 0
-                    else
-                        ${increment} : ${checkHits}
-                `;
-                //coerce the expression into single-line
-                trackingExpression = trackingExpression.replace(/\n/gi, '').replace(/\s+/g, ' ').trim();
-                // Add the tracking expression right before this line
-                lines.push(new SourceNode(breakpoint.line, 0, originalFilePath, trackingExpression));
-            }
-        } else {
-            // add a STOP statement right before this line. Map the stop code to the line the breakpoint represents
-            //because otherwise source-map will return null for this location
-            lines.push(new SourceNode(breakpoint.line, 0, originalFilePath, 'STOP'));
-        }
-        return lines;
+    public getBreakpointsForFile(filePath: string): AugmentedSourceBreakpoint[] {
+        let sanitizedPath = this.sanitizeSourceFilePath(filePath);
+        return this.breakpointsByFilePath[sanitizedPath] ?? [];
     }
 
     /**
      * Get the list of breakpoints for the specified file path, or an empty array
      */
-    public getBreakpointsForFile(filePath: string): AugmentedSourceBreakpoint[] {
-        let key = this.sanitizeSourceFilePath(filePath);
-        return this.breakpointsByFilePath[key] ?? [];
+    public removeQueuedBreakpointsForFile(filePath: string) {
+        let sanitizedPath = this.sanitizeSourceFilePath(filePath);
+        delete this.breakpointsQueue[sanitizedPath];
     }
 
     /**
@@ -421,6 +338,8 @@ interface AugmentedSourceBreakpoint extends DebugProtocol.SourceBreakpoint {
      * breakpoints, we need to mark them as hidden when the user unsets them.
      */
     isHidden: boolean;
+
+    protocolBreakpointIds: number[];
 }
 
 interface BreakpointWorkItem {
